moving files and public entrypoint
