+++ application/src/main/java/com/cleancrud/spark/entrypoint/GetEntryPoint.java
@@ -1,6 +1,5 @@
-package com.cleancrud.spark.entrypoint.impl;
+package com.cleancrud.spark.entrypoint;
 
-import com.cleancrud.spark.entrypoint.EntryPoint;
 import com.cleancrud.spark.utils.JsonTransformer;
 import java.util.HashMap;
 import java.util.Map;
@@ -10,27 +9,28 @@ import spark.Request;
 import spark.Response;
 
 /**
- * DELETE entrypoint.
+ * GET entrypoint.
  */
 @Singleton
-public class DeleteEntryPoint extends EntryPoint {
+public class GetEntryPoint extends EntryPoint {
 
   @Inject
-  public DeleteEntryPoint(JsonTransformer json) {
+  public GetEntryPoint(JsonTransformer json) {
     super(json);
   }
 
   @Override
-  protected Response internalHandle(Request request, Response response) {
+  public Response internalHandle(Request request, Response response) {
     Map<String, String> result = new HashMap<>();
 
     try {
       // TODO here we use the usecases from domain
-      result.put("result", "DELETE RESPONSE");
+      result.put("result", "GET RESPONSE");
       response.body(serialize(result));
     } catch (Exception e) {
       e.printStackTrace();
     }
+
     return response;
   }
 }
