+++ application/src/main/java/com/cleancrud/spark/module/EntryPointModule.java
@@ -1,9 +1,9 @@
 package com.cleancrud.spark.module;
 
-import com.cleancrud.spark.entrypoint.impl.DeleteEntryPoint;
-import com.cleancrud.spark.entrypoint.impl.GetEntryPoint;
-import com.cleancrud.spark.entrypoint.impl.PostEntryPoint;
-import com.cleancrud.spark.entrypoint.impl.PutEntryPoint;
+import com.cleancrud.spark.entrypoint.DeleteEntryPoint;
+import com.cleancrud.spark.entrypoint.GetEntryPoint;
+import com.cleancrud.spark.entrypoint.PostEntryPoint;
+import com.cleancrud.spark.entrypoint.PutEntryPoint;
 import com.google.inject.AbstractModule;
 
 public class EntryPointModule extends AbstractModule {
